% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastcore.R
\name{.rc_build_meta_module_gem_core}
\alias{.rc_build_meta_module_gem_core}
\title{Build one GRN-defined meta-module GEM with add-only FASTCORE completion}
\usage{
.rc_build_meta_module_gem_core(
  gem,
  reaction_membership,
  core_reactions = NULL,
  sample_id,
  module_id,
  medium_table = NULL,
  condition = NULL,
  target_direction = c("both", "forward", "reverse"),
  solver = "highs",
  time_limit = 300,
  fastcore_epsilon = 1e-04,
  max_support_reactions = 2000,
  strict = TRUE
)
}
\description{
Build one GRN-defined meta-module GEM with add-only FASTCORE completion
}
