% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pando_grn.R
\name{.rc_signed_relation}
\alias{.rc_signed_relation}
\title{Run sample-specific Pando GRNs and construct reaction meta-modules}
\usage{
.rc_signed_relation(values)
}
\description{
Run sample-specific Pando GRNs and construct reaction meta-modules
}
