% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metacell.R
\name{rc_aggregate_fragments_by_membership}
\alias{rc_aggregate_fragments_by_membership}
\title{Aggregate ATAC fragments by metacell membership}
\usage{
rc_aggregate_fragments_by_membership(
  fragment_files,
  membership,
  outdir,
  tmp_root = tempdir(),
  bgzip_path = "bgzip",
  tabix_path = "tabix",
  nb_cl = 1L
)
}
\description{
Aggregate ATAC fragments by metacell membership
}
