% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudobulk.R
\name{rc_atac_unit_logcpm}
\alias{rc_atac_unit_logcpm}
\title{Filter ATAC peaks detected in at least min_pools and compute pool logCPM}
\usage{
rc_atac_unit_logcpm(atac_counts, unit_map, min_pools = 3, BPPARAM = NULL)
}
\description{
Filter ATAC peaks detected in at least min_pools and compute pool logCPM
}
