% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/full_gem.R
\name{rc_build_full_gem_cache}
\alias{rc_build_full_gem_cache}
\title{Cache one complete full GEM per medium scenario}
\usage{
rc_build_full_gem_cache(
  gem,
  dirs,
  medium_scenarios,
  cache_dir = tempfile("RegCompassR_full_gem_cache_"),
  force = FALSE,
  conditions = NULL
)
}
\description{
Cache one complete full GEM per medium scenario
}
