% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{rc_describe_microcompass_by_group}
\alias{rc_describe_microcompass_by_group}
\title{Test sample-level microCOMPASS differential scores}
\usage{
rc_describe_microcompass_by_group(
  result,
  sample_col = "sample_id",
  condition_col = "condition",
  celltype_col = "cell_type"
)
}
\description{
Test sample-level microCOMPASS differential scores
}
