% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{rc_export_microcompass}
\alias{rc_export_microcompass}
\title{Export standardized microCOMPASS outputs}
\usage{
rc_export_microcompass(
  result,
  outdir,
  write_matrices = TRUE,
  write_diagnostics = TRUE
)
}
\description{
Export standardized microCOMPASS outputs
}
