% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pando_grn.R
\name{rc_extract_pando_tf_peak_gene}
\alias{rc_extract_pando_tf_peak_gene}
\title{Extract and filter a Pando TF-peak-gene coefficient table}
\usage{
rc_extract_pando_tf_peak_gene(
  grn_object,
  sample_id,
  padj_threshold = 0.05,
  min_abs_estimate = 0,
  min_model_rsq = 0.1,
  require_padj = TRUE
)
}
\description{
Extract and filter a Pando TF-peak-gene coefficient table
}
