% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudobulk.R
\name{rc_filter_empty_units}
\alias{rc_filter_empty_units}
\title{Filter empty pseudobulk pools before normalization}
\usage{
rc_filter_empty_units(pb_counts, unit_meta)
}
\description{
Filter empty pseudobulk pools before normalization
}
