% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpr_capacity.R
\name{rc_gene_score}
\alias{rc_gene_score}
\title{Gene capacity score from meta-cell-level logCPM}
\usage{
rc_gene_score(
  X,
  min_scale = 0.05,
  z_clip = 6,
  mode = c("absolute", "relative"),
  half_saturation = getOption("RegCompassR.cpm_half_saturation", 1)
)
}
\description{
Gene capacity score from meta-cell-level logCPM
}
