% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metacell.R
\name{rc_import_supercell2_metacells}
\alias{rc_import_supercell2_metacells}
\title{Import saved SuperCell2.0 metacell outputs}
\usage{
rc_import_supercell2_metacells(
  metacell_dirs,
  rna_assay = "RNA",
  atac_assay = "ATAC",
  sample_col = "sample_id",
  condition_col = "condition",
  celltype_col = "cell_type",
  require_fragments = FALSE
)
}
\description{
Import saved SuperCell2.0 metacell outputs
}
