% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metacell.R
\name{rc_load_or_merge_metacell_objects}
\alias{rc_load_or_merge_metacell_objects}
\title{Load and merge saved metacell Seurat objects}
\usage{
rc_load_or_merge_metacell_objects(
  metacell_objects,
  fragment_manifest = NULL,
  metacell_meta = NULL,
  fragment_files = NULL,
  rna_assay = "RNA",
  atac_assay = "ATAC",
  require_complete_fragments = TRUE
)
}
\description{
Load and merge saved metacell Seurat objects
}
