% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gem_io.R
\name{rc_make_gem}
\alias{rc_make_gem}
\title{Construct a minimal GEM object from stoichiometry and bounds}
\usage{
rc_make_gem(
  S,
  lb = NULL,
  ub = NULL,
  reaction_meta = NULL,
  metabolite_meta = NULL,
  medium_policy = "base_bounds",
  model_info = NULL
)
}
\description{
Construct a minimal GEM object from stoichiometry and bounds
}
