% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_module_definition.R
\name{rc_map_meta_module_core_reactions}
\alias{rc_map_meta_module_core_reactions}
\title{Map GRN metabolic gene modules to core Human-GEM reactions}
\usage{
rc_map_meta_module_core_reactions(gene_nodes, gpr_table)
}
\description{
Map GRN metabolic gene modules to core Human-GEM reactions
}
