% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpr_parser.R
\name{rc_parse_gpr_table}
\alias{rc_parse_gpr_table}
\title{Parse a reaction GPR table}
\usage{
rc_parse_gpr_table(gpr_table)
}
\description{
Accepts either \code{reaction_id + gpr} or a pre-parsed long table with
\code{reaction_id + and_group_id + gene}.
}
