% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpr_parser.R
\name{rc_promiscuity_weight}
\alias{rc_promiscuity_weight}
\title{Compute gene promiscuity weights from parsed GPR rules}
\usage{
rc_promiscuity_weight(gpr_list, mode = c("sqrt", "linear", "none"))
}
\description{
Compute gene promiscuity weights from parsed GPR rules
}
