% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpr_capacity.R
\name{rc_reaction_capacity_one}
\alias{rc_reaction_capacity_one}
\title{Compute capacity for one reaction in one pool}
\usage{
rc_reaction_capacity_one(
  parsed_gpr,
  gene_score_vec,
  tau = 0.2,
  and_method = c("min", "boltzmann", "mean"),
  or_method = c("max", "sum_sqrtK", "prob_or", "sum")
)
}
\description{
Compute capacity for one reaction in one pool
}
