% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudobulk.R
\name{rc_unit_bulk_counts}
\alias{rc_unit_bulk_counts}
\title{Pseudobulk raw counts by pool}
\usage{
rc_unit_bulk_counts(counts, unit_map, fun = c("sum", "mean"), BPPARAM = NULL)
}
\description{
Pseudobulk raw counts by pool
}
