% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudobulk.R
\name{rc_unit_detection}
\alias{rc_unit_detection}
\title{Compute meta-cell-level detection rates}
\usage{
rc_unit_detection(counts, unit_map, BPPARAM = NULL)
}
\description{
Compute meta-cell-level detection rates
}
