% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metacell.R
\name{rc_validate_metacell_inputs}
\alias{rc_validate_metacell_inputs}
\title{Validate metacell-level RegCompass inputs}
\usage{
rc_validate_metacell_inputs(
  rna_metacell_counts,
  metacell_meta,
  atac_metacell_counts = NULL,
  metacell_id_col = "metacell_id",
  sample_col = "sample_id",
  condition_col = "condition",
  celltype_col = "cell_type"
)
}
\description{
Validate metacell-level RegCompass inputs
}
